pyproject.toml
[project]
name = "substrate"
version = "0.1.0"
description = "Personal business operating system framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.0.0",
    "psycopg-pool>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic-ai>=0.0.20",
    "pyyaml>=6.0",
    "python-slugify>=8.0.0",
    "python-dotenv>=1.0.0",
    "absurd-sdk",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["substrate"]

[tool.uv.sources]
absurd-sdk = { path = "libs/absurd-sdk" }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
